struct: change strcture to match dev branch
