--- hr-finder/src/HRmanager0228/cli/controller/EmployeeController.java
@@ -1,4 +1,0 @@
-package HRmanager0228.cli.controller;
-
-public class EmployeeController {
-}
