--- hr-finder/src/HRmanager0228/dao/EmployeeDao.java
@@ -1,4 +1,0 @@
-package HRmanager0228.dao;
-
-public interface EmployeeDao {
-}
