--- hr-finder/src/HRmanager0228/dao/SalaryDaoImpl.java
@@ -1,4 +1,0 @@
-package HRmanager0228.dao;
-
-public class SalaryDaoImpl {
-}
