--- hr-finder/src/HRmanager0228/util/utildemo.java
@@ -1,4 +1,0 @@
-package HRmanager0228.util;
-// validator같은 공통된 기능을 제공하는.. 예외처리도 ?
-public class utildemo {
-}
