+++ hr-finder/src/cli/controller/EmployeeController.java
@@ -1,0 +1,4 @@
+package cli.controller;
+
+public class EmployeeController {
+}
