+++ hr-finder/src/cli/controller/SalaryController.java
@@ -1,0 +1,4 @@
+package cli.controller;
+
+public class SalaryController {
+}
