+++ hr-finder/src/cli/io/EmployeeIO.java
@@ -1,0 +1,4 @@
+package cli.io;
+
+public class EmployeeIO {
+}
