+++ hr-finder/src/config/configdemo.java
@@ -1,0 +1,4 @@
+package config;
+// db나 파일 연결,로드 등 관련 환경 통합관리
+public class configdemo {
+}
