+++ hr-finder/src/dao/EmployeeDao.java
@@ -1,0 +1,4 @@
+package dao;
+
+public interface EmployeeDao {
+}
