+++ hr-finder/src/dao/EmployeeDaoImpl.java
@@ -1,0 +1,4 @@
+package dao;
+
+public class EmployeeDaoImpl {
+}
