+++ hr-finder/src/dao/SalaryDao.java
@@ -1,0 +1,4 @@
+package dao;
+
+public interface SalaryDao {
+}
