+++ hr-finder/src/dao/SalaryDaoImpl.java
@@ -1,0 +1,4 @@
+package dao;
+
+public class SalaryDaoImpl {
+}
