+++ hr-finder/src/docs/Dev_guide.md
