+++ hr-finder/src/service/EmployeeService.java
@@ -1,0 +1,4 @@
+package service;
+
+public interface EmployeeService {
+}
