+++ hr-finder/src/service/EmployeeServiceImpl.java
@@ -1,0 +1,4 @@
+package service;
+
+public class EmployeeServiceImpl {
+}
