+++ hr-finder/src/util/utildemo.java
@@ -1,0 +1,4 @@
+package util;
+// validator같은 공통된 기능을 제공하는.. 예외처리도 ?
+public class utildemo {
+}
